Update merkletree, go-scale and post dependencies (#3672)
## Motivation
Closes https://github.com/spacemeshos/pm/issues/125 for go-spacemesh

## Changes
Update dependencies and fix types where necessary

## Test Plan
Existing tests pass

## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)

+++ activation/handler.go
@@ -193,7 +193,7 @@ func (h *Handler) SyntacticallyValidateAtx(ctx context.Context, atx *types.Activ
 		return nil, fmt.Errorf("validation failed: initial atx not found: %w", err)
 	}
 
-	leaves, err := h.nipostValidator.Validate(commitment, atx.NIPost, *expectedChallengeHash, uint(atx.NumUnits))
+	leaves, err := h.nipostValidator.Validate(commitment, atx.NIPost, *expectedChallengeHash, atx.NumUnits)
 	if err != nil {
 		return nil, fmt.Errorf("invalid nipost: %w", err)
 	}
@@ -242,7 +242,7 @@ func (h *Handler) validateInitialAtx(ctx context.Context, atx *types.ActivationT
 	initialPostMetadata.Challenge = shared.ZeroChallenge
 
 	commitment := GetCommitmentBytes(atx.NodeID(), *atx.CommitmentATX)
-	if err := h.nipostValidator.ValidatePost(commitment, atx.InitialPost, &initialPostMetadata, uint(atx.NumUnits)); err != nil {
+	if err := h.nipostValidator.ValidatePost(commitment, atx.InitialPost, &initialPostMetadata, atx.NumUnits); err != nil {
 		return fmt.Errorf("invalid initial Post: %w", err)
 	}
